pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "caption-agent"
version = "0.1.0"
description = "MVP caption agent for person instance attributes and relations."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "pydantic>=2.7.0",
  "langchain>=1.0.0",
  "langchain-openai>=0.3.0",
  "langgraph>=1.0.0",
  "pillow>=10.0.0",
  "pyyaml>=6.0.0",
  "python-dotenv>=1.0.0",
  "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
